Move dev server specific stuff to own file
Refs #34.
